# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's Moon Group Transparency",
        "description": "API's for transparency of Moon Group",
        "version": "1.0.0"
    },
    "paths": {
...
        "/moontransparency/public/api/Indicator": {
            "get": {
                "tags": [
                    "Indicator"
                ],
                "summary": "Obtener información de Indicators con filtros y ordenamiento",
                "operationId": "91c2ffd1bb2530ac4d9ec7735c18fd8c",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre de Indicator",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "Filtrar por ID del proyecto",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "indicator_name",
                        "in": "query",
                        "description": "Filtrar por nombre del indicador",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "target_value",
                        "in": "query",
                        "description": "Filtrar por valor objetivo",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "progress_value",
                        "in": "query",
                        "description": "Filtrar por valor de progreso",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "unit",
                        "in": "query",
                        "description": "Filtrar por unidad del indicador",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Fecha de inicio",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Fecha de fin",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de Indicators",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Indicator"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Indicator"
                ],
                "summary": "Crear Indicator",
                "operationId": "ea79a0d6bbea3edc062d4255d2f0d6d2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/IndicatorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Donación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Indicator"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "La validación falló."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/moontransparency/public/api/Indicator/{id}": {
            "get": {
                "tags": [
                    "Indicator"
                ],
                "summary": "Obtener detalles de un Indicator por ID",
                "operationId": "f0ab7424e78ee4df5bd3b11d8dcc3805",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del Indicator",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donación encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Indicator"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donación no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Donación no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Indicator"
                ],
                "summary": "Actualizar un Indicator",
                "operationId": "8417e1386bb7f14d840c0ada973163fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/IndicatorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Donación actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Indicator"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Datos inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donación no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Donación no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Indicator"
                ],
                "summary": "Eliminar un Indicator por ID",
                "operationId": "717c5af4688ecc38f9c217f29b586a19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donación eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Donación eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Donación no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "IndicatorRequest": {
                "required": [
                    "project_id",
                    "indicator_name",
                    "target_value",
                    "progress_value",
                    "unit"
                ],
                "properties": {
                    "project_id": {
                        "type": "integer",
                        "example": "101"
                    },
                    "indicator_name": {
                        "type": "string",
                        "example": "Energía Renovable"
                    },
                    "target_value": {
                        "type": "number",
                        "format": "float",
                        "example": "1000000"
                    },
                    "progress_value": {
                        "type": "number",
                        "format": "float",
                        "example": "500000"
                    },
                    "unit": {
                        "type": "string",
                        "example": "KWh"
                    }
                },
                "type": "object"
            },
            "Indicator": {
                "title": "Indicator",
                "description": "Modelo de Indicator",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "project_id": {
                        "type": "integer",
                        "example": "101"
                    },
                    "indicator_name": {
                        "type": "string",
                        "example": "Energía Renovable"
                    },
                    "target_value": {
                        "type": "number",
                        "format": "float",
                        "example": "1000000"
                    },
                    "progress_value": {
                        "type": "number",
                        "format": "float",
                        "example": "500000"
                    },
                    "unit": {
                        "type": "string",
                        "example": "KWh"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-26T21:44:24"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}